firestore-pgvector/pyproject.toml
[tool.poetry]
name = "firestore-pgvector"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "functions"}]

[tool.poetry.dependencies]
python = "^3.11"
firebase-functions = "^0.1.1"
firebase-admin = "^6.2.0"
google-cloud-aiplatform = "^1.28.0"
pgvector = "^0.1.8"
cloud-sql-python-connector = "^1.3.0"
asyncpg = "^0.28.0"
google-cloud-tasks = "^2.13.2"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.21.0"
black = "^23.7.0"

[tool.poe.tasks.prepare]
script = "scripts:prepare_for_deployment"
help = "Prepare for deployment"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
